Изменения в task_21

+++ task_21/Program.cs
@@ -5,16 +5,16 @@
 
 Console.Clear();
 Console.Write("введите координаты первой точки x1 = ");
-int x1 = Convert.ToInt32(Console.ReadLine());
+double x1 = Convert.ToDouble(Console.ReadLine());
 Console.Write("введите координаты первой точки y1 = ");
-int y1 = Convert.ToInt32(Console.ReadLine());
+double y1 = Convert.ToDouble(Console.ReadLine());
 Console.Write("введите координаты первой точки z1 = ");
-int z1 = Convert.ToInt32(Console.ReadLine());
+double z1 = Convert.ToDouble(Console.ReadLine());
 Console.Write("введите координаты второй точки x2 = ");
-int x2 = Convert.ToInt32(Console.ReadLine());
+double x2 = Convert.ToDouble(Console.ReadLine());
 Console.Write("введите координаты второй точки y2 = ");
-int y2 = Convert.ToInt32(Console.ReadLine());
+double y2 = Convert.ToDouble(Console.ReadLine());
 Console.Write("введите координаты второй точки z2 = ");
-int z2 = Convert.ToInt32(Console.ReadLine());
+double z2 = Convert.ToDouble(Console.ReadLine());
 
 Console.Write(Math.Sqrt(Math.Pow(x1 - x2, 2) + Math.Pow(y2 - y2, 2) + Math.Pow(z1 - z2, 2)));
